Removed backward compatibility of Signer and StringSigner
